Move products data layer handlers to state/sites/products/handlers.js
