+++ client/extensions/woocommerce/state/sites/products/handlers.js
